OP-211 fix decoders

See merge request open-platform/api!179

+++ src/main/kotlin/io/openfuture/api/domain/event/PaidForShareHolderEvent.kt
@@ -2,11 +2,11 @@ package io.openfuture.api.domain.event
 
 import io.openfuture.api.component.web3.event.EventType
 import io.openfuture.api.component.web3.event.EventType.PAID_FOR_SHARE_HOLDER
-import java.math.BigInteger
+import java.math.BigDecimal
 
 data class PaidForShareHolderEvent(
         val userAddress: String,
-        val amount: BigInteger
+        val amount: BigDecimal
 ) : Event {
 
     override fun getType(): EventType = PAID_FOR_SHARE_HOLDER
